+++ axum-core/src/lib.rs
@@ -31,3 +31,6 @@ pub mod response;
 pub type BoxError = Box<dyn std::error::Error + Send + Sync>;
 
 pub use self::ext_traits::{request::RequestExt, request_parts::RequestPartsExt};
+
+#[cfg(test)]
+use axum_macros::__private_axum_test as test;
